♻️ move UniqueMessage to common infrastructure
