+++ backend/api/src/application/project/remove_leader.rs
@@ -1,10 +1,10 @@
 use std::sync::Arc;
 
 use domain::{
-	AggregateRootRepository, Destination, DomainError, Event, Project, ProjectId, Publisher,
-	UniqueMessage, UserId,
+	AggregateRootRepository, Destination, DomainError, Event, Project, ProjectId, Publisher, UserId,
 };
 use event_store::bus::QUEUE_NAME as EVENT_STORE_QUEUE;
+use infrastructure::amqp::UniqueMessage;
 
 pub struct Usecase {
 	event_publisher: Arc<dyn Publisher<UniqueMessage<Event>>>,
