+++ backend/common/domain/src/lib.rs
@@ -16,7 +16,7 @@ pub use error::*;
 mod messaging;
 pub use messaging::{
 	Destination, Message, Payload as MessagePayload, Publisher, PublisherError, Subscriber,
-	SubscriberCallbackError, SubscriberError, UniqueMessage,
+	SubscriberCallbackError, SubscriberError,
 };
 
 mod project;
