+++ backend/common/infrastructure/src/amqp/mod.rs
@@ -6,3 +6,6 @@ pub use config::Config;
 
 mod publisher;
 mod subscriber;
+
+mod unique_message;
+pub use unique_message::UniqueMessage;
